pyproject.toml
[tool.poetry]
name = "urbanaut-bot"
version = "0.0.1"
description = ""
authors = ["crazyproger1 <crazyproger1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.8.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.10.1"
isort = "^5.13.2"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
